+++ src/collective/recipe/template/tests.py
@@ -13,13 +13,15 @@ def setUp(test):
     zc.buildout.testing.buildoutSetUp(test)
     zc.buildout.testing.install_develop('collective.recipe.template', test)
     zc.buildout.testing.install('zope.testing', test)
+    zc.buildout.testing.install('six', test)
     zc.buildout.testing.install('Genshi', test)
 
 
 checker = renormalizing.RENormalizing([
     zc.buildout.testing.normalize_path,
     (re.compile('#![^\n]+\n'), ''),
-    (re.compile('-\S+-py\d[.]\d(-\S+)?.egg'), '-pyN.N.egg'),
+    (re.compile(r'-\S+-py\d[.]\d(-\S+)?.egg'), '-pyN.N.egg',
+     ),
 ])
 
 
@@ -27,18 +29,12 @@ def test_suite():
     return unittest.TestSuite([
         doctest.DocFileSuite(
             'README.rst',
-            setUp=setUp,
-            tearDown=zc.buildout.testing.buildoutTearDown,
+            setUp=setUp, tearDown=zc.buildout.testing.buildoutTearDown,
             optionflags=doctest.ELLIPSIS | doctest.NORMALIZE_WHITESPACE,
             checker=checker),
         doctest.DocFileSuite(
             'genshitemplate.rst',
-            setUp=setUp,
-            tearDown=zc.buildout.testing.buildoutTearDown,
+            setUp=setUp, tearDown=zc.buildout.testing.buildoutTearDown,
             optionflags=doctest.ELLIPSIS | doctest.NORMALIZE_WHITESPACE,
             checker=checker),
-        ])
-
-
-if __name__ == '__main__':
-    unittest.main(defaultTest='test_suite')
+    ])